// src/app/srvapi/models/processor-operation-desc.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ProcessorOperationDesc
 */
export interface ProcessorOperationDesc {
    /**
     * 
     * @type {Array&lt;ProcessorOperationArgumentDesc&gt;}
     * @memberof ProcessorOperationDesc
     */
    arguments?: any;
    /**
     * 
     * @type {string}
     * @memberof ProcessorOperationDesc
     */
    name?: any;
    /**
     * 
     * @type {string}
     * @memberof ProcessorOperationDesc
     */
    result?: ProcessorOperationDescResultEnum;
}

/**
    * @export
    * @enum {string}
    */
export enum ProcessorOperationDescResultEnum {
    FLOAT = 'FLOAT',
    INT = 'INT',
    STRING = 'STRING',
    DATE = 'DATE',
    FLOATLIST = 'FLOAT_LIST',
    INTLIST = 'INT_LIST',
    STRINGLIST = 'STRING_LIST',
    DATELIST = 'DATE_LIST',
    VOID = 'VOID'
}